Move tx_msg under core
