+++ crates/ibc/src/applications/transfer/msgs/transfer.rs
@@ -11,7 +11,7 @@ use crate::applications::transfer::error::TokenTransferError;
 use crate::core::ics04_channel::timeout::TimeoutHeight;
 use crate::core::ics24_host::identifier::{ChannelId, PortId};
 use crate::core::timestamp::Timestamp;
-use crate::tx_msg::Msg;
+use crate::core::tx_msg::Msg;
 
 pub const TYPE_URL: &str = "/ibc.applications.transfer.v1.MsgTransfer";
