+++ crates/ibc/src/core/ics02_client/msgs/create_client.rs
@@ -7,8 +7,8 @@ use ibc_proto::ibc::core::client::v1::MsgCreateClient as RawMsgCreateClient;
 use ibc_proto::protobuf::Protobuf;
 
 use crate::core::ics02_client::error::ClientError;
+use crate::core::tx_msg::Msg;
 use crate::signer::Signer;
-use crate::tx_msg::Msg;
 
 pub const TYPE_URL: &str = "/ibc.core.client.v1.MsgCreateClient";
