+++ crates/ibc/src/core/ics03_connection/msgs/conn_open_init.rs
@@ -9,8 +9,8 @@ use crate::core::ics03_connection::connection::Counterparty;
 use crate::core::ics03_connection::error::ConnectionError;
 use crate::core::ics03_connection::version::Version;
 use crate::core::ics24_host::identifier::ClientId;
+use crate::core::tx_msg::Msg;
 use crate::signer::Signer;
-use crate::tx_msg::Msg;
 
 pub const TYPE_URL: &str = "/ibc.core.connection.v1.MsgConnectionOpenInit";
