+++ crates/ibc/src/core/ics04_channel/msgs/chan_close_init.rs
@@ -6,8 +6,8 @@ use ibc_proto::ibc::core::channel::v1::MsgChannelCloseInit as RawMsgChannelClose
 
 use crate::core::ics04_channel::error::ChannelError;
 use crate::core::ics24_host::identifier::{ChannelId, PortId};
+use crate::core::tx_msg::Msg;
 use crate::signer::Signer;
-use crate::tx_msg::Msg;
 
 pub const TYPE_URL: &str = "/ibc.core.channel.v1.MsgChannelCloseInit";
