+++ crates/ibc/src/core/ics04_channel/msgs/chan_open_init.rs
@@ -4,9 +4,9 @@ use crate::core::ics04_channel::channel::{Order, State};
 use crate::core::ics04_channel::error::ChannelError;
 use crate::core::ics04_channel::Version;
 use crate::core::ics24_host::identifier::{ConnectionId, PortId};
+use crate::core::tx_msg::Msg;
 use crate::prelude::*;
 use crate::signer::Signer;
-use crate::tx_msg::Msg;
 
 use ibc_proto::ibc::core::channel::v1::MsgChannelOpenInit as RawMsgChannelOpenInit;
 use ibc_proto::protobuf::Protobuf;
